Development

+++ lib/cuckoo/core/scheduler.py
@@ -1,30 +1,44 @@
-# Copyright (C) 2010-2013 Cuckoo Sandbox Developers.
+# Copyright (C) 2010-2014 Cuckoo Sandbox Developers.
 # This file is part of Cuckoo Sandbox - http://www.cuckoosandbox.org
 # See the file 'docs/LICENSE' for copying permission.
 
 import os
-import sys
 import time
 import shutil
 import logging
 import Queue
 from threading import Thread, Lock
 
-from lib.cuckoo.common.constants import CUCKOO_ROOT
-from lib.cuckoo.common.exceptions import CuckooMachineError, CuckooGuestError, CuckooOperationalError, CuckooCriticalError
-from lib.cuckoo.common.abstracts import  Machinery
-from lib.cuckoo.common.objects import Dictionary, File
-from lib.cuckoo.common.utils import  create_folders, create_folder
 from lib.cuckoo.common.config import Config
+from lib.cuckoo.common.constants import CUCKOO_ROOT
+from lib.cuckoo.common.exceptions import CuckooMachineError, CuckooGuestError
+from lib.cuckoo.common.exceptions import CuckooOperationalError
+from lib.cuckoo.common.exceptions import CuckooCriticalError
+from lib.cuckoo.common.objects import File
+from lib.cuckoo.common.utils import create_folder
 from lib.cuckoo.core.database import Database, TASK_COMPLETED, TASK_REPORTED
 from lib.cuckoo.core.guest import GuestManager
+from lib.cuckoo.core.plugins import list_plugins, RunAuxiliary, RunProcessing
+from lib.cuckoo.core.plugins import RunSignatures, RunReporting
 from lib.cuckoo.core.resultserver import Resultserver
-from lib.cuckoo.core.plugins import import_plugin, list_plugins, RunAuxiliary, RunProcessing, RunSignatures, RunReporting
 
 log = logging.getLogger(__name__)
 
 machinery = None
 machine_lock = Lock()
+
+total_analysis_count = 0
+active_analysis_count = 0
+
+
+class CuckooDeadMachine(Exception):
+    """Exception thrown when a machine turns dead.
+
+    When this exception has been thrown, the analysis task will start again,
+    and will try to use another machine, when available.
+    """
+    pass
+
 
 class AnalysisManager(Thread):
     """Analysis Manager.
@@ -71,6 +85,17 @@ class AnalysisManager(Thread):
 
         return True
 
+    def check_file(self):
+        """Checks the integrity of the file to be analyzed."""
+        sample = Database().view_sample(self.task.sample_id)
+
+        sha256 = File(self.task.target).get_sha256()
+        if sha256 != sample.sha256:
+            log.error("Target file has been modified after submission: \"%s\"", self.task.target)
+            return False
+
+        return True
+
     def store_file(self):
         """Store a copy of the file being analyzed."""
         if not os.path.exists(self.task.target):
@@ -101,7 +126,8 @@ class AnalysisManager(Thread):
             else:
                 shutil.copy(self.binary, new_binary_path)
         except (AttributeError, OSError) as e:
-            log.error("Unable to create symlink/copy from \"%s\" to \"%s\"", self.binary, self.storage)
+            log.error("Unable to create symlink/copy from \"%s\" to "
+                      "\"%s\": %s", self.binary, self.storage, e)
 
         return True
 
@@ -112,11 +138,23 @@ class AnalysisManager(Thread):
         # Start a loop to acquire the a machine to run the analysis on.
         while True:
             machine_lock.acquire()
-            # If the user specified a specific machine ID or a platform to be
-            # used, acquire the machine accordingly.
-            machine = machinery.acquire(machine_id=self.task.machine,
-                                        platform=self.task.platform)
-            machine_lock.release()
+
+            # In some cases it's possible that we enter this loop without
+            # having any available machines. We should make sure this is not
+            # such case, or the analysis task will fail completely.
+            if not machinery.availables():
+                machine_lock.release()
+                time.sleep(1)
+                continue
+
+            # If the user specified a specific machine ID, a platform to be
+            # used or machine tags acquire the machine accordingly.
+            try:
+                machine = machinery.acquire(machine_id=self.task.machine,
+                                            platform=self.task.platform,
+                                            tags=self.task.tags)
+            finally:
+                machine_lock.release()
 
             # If no machine is available at this moment, wait for one second
             # and try again.
@@ -124,7 +162,8 @@ class AnalysisManager(Thread):
                 log.debug("Task #%d: no machine available yet", self.task.id)
                 time.sleep(1)
             else:
-                log.info("Task #%d: acquired machine %s (label=%s)", self.task.id, machine.name, machine.label)
+                log.info("Task #%d: acquired machine %s (label=%s)",
+                         self.task.id, machine.name, machine.label)
                 break
 
         self.machine = machine
@@ -159,25 +198,36 @@ class AnalysisManager(Thread):
     def launch_analysis(self):
         """Start analysis."""
         succeeded = False
+        dead_machine = False
 
-        log.info("Starting analysis of %s \"%s\" (task=%d)", self.task.category.upper(), self.task.target, self.task.id)
+        log.info("Starting analysis of %s \"%s\" (task=%d)",
+                 self.task.category.upper(), self.task.target, self.task.id)
 
         # Initialize the the analysis folders.
         if not self.init_storage():
             return False
 
         if self.task.category == "file":
+            # Check whether the file has been changed for some unknown reason.
+            # And fail this analysis if it has been modified.
+            if not self.check_file():
+                return False
+
             # Store a copy of the original file.
             if not self.store_file():
                 return False
 
         # Acquire analysis machine.
-        self.acquire_machine()
+        try:
+            self.acquire_machine()
+        except CuckooOperationalError as e:
+            log.error("Cannot acquire machine: {0}".format(e))
+            return False
 
         # Generate the analysis configuration file.
         options = self.build_options()
 
-        # At this point we can tell the Resultserver about it
+        # At this point we can tell the Resultserver about it.
         try:
             Resultserver().add_task(self.task, self.machine)
         except Exception as e:
@@ -196,12 +246,8 @@ class AnalysisManager(Thread):
             # Start the machine.
             machinery.start(self.machine.label)
         except CuckooMachineError as e:
-            log.error(str(e), extra={"task_id" : self.task.id})
-
-            # Stop Auxiliary modules.
-            aux.stop()
-
-            return False
+            log.error(str(e), extra={"task_id": self.task.id})
+            dead_machine = True
         else:
             try:
                 # Initialize the guest manager.
@@ -209,19 +255,14 @@ class AnalysisManager(Thread):
                 # Start the analysis.
                 guest.start_analysis(options)
             except CuckooGuestError as e:
-                log.error(str(e), extra={"task_id" : self.task.id})
-
-                # Stop Auxiliary modules.
-                aux.stop()
-
-                return False
+                log.error(str(e), extra={"task_id": self.task.id})
             else:
                 # Wait for analysis completion.
                 try:
                     guest.wait_for_completion()
                     succeeded = True
                 except CuckooGuestError as e:
-                    log.error(str(e), extra={"task_id" : self.task.id})
+                    log.error(str(e), extra={"task_id": self.task.id})
                     succeeded = False
 
         finally:
@@ -232,10 +273,10 @@ class AnalysisManager(Thread):
             if self.cfg.cuckoo.memory_dump or self.task.memory:
                 try:
                     machinery.dump_memory(self.machine.label,
-                                         os.path.join(self.storage, "memory.dmp"))
+                                          os.path.join(self.storage, "memory.dmp"))
                 except NotImplementedError:
                     log.error("The memory dump functionality is not available "
-                              "for current machine manager")
+                              "for the current machine manager")
                 except CuckooMachineError as e:
                     log.error(e)
 
@@ -243,20 +284,41 @@ class AnalysisManager(Thread):
                 # Stop the analysis machine.
                 machinery.stop(self.machine.label)
             except CuckooMachineError as e:
-                log.warning("Unable to stop machine %s: %s", self.machine.label, e)
+                log.warning("Unable to stop machine %s: %s",
+                            self.machine.label, e)
 
-            # Market the machine in the database as stopped.
+            # Mark the machine in the database as stopped. Unless this machine
+            # has been marked as dead, we just keep it as "started" in the
+            # database so it'll not be used later on in this session.
             Database().guest_stop(guest_log)
 
+            # After all this, we can make the Resultserver forget about the
+            # internal state for this analysis task.
+            Resultserver().del_task(self.task, self.machine)
+
+            if dead_machine:
+                # Remove the guest from the database, so that we can assign a
+                # new guest when the task is being analyzed with another
+                # machine.
+                Database().guest_remove(guest_log)
+
+                # Remove the analysis directory that has been created so
+                # far, as launch_analysis() is going to be doing that again.
+                shutil.rmtree(self.storage)
+
+                # This machine has turned dead, so we throw an exception here
+                # which informs the AnalysisManager that it should analyze
+                # this task again with another available machine.
+                raise CuckooDeadMachine()
+
             try:
-                # Release the analysis machine.
+                # Release the analysis machine. But only if the machine has
+                # not turned dead yet.
                 machinery.release(self.machine.label)
             except CuckooMachineError as e:
                 log.error("Unable to release machine %s, reason %s. "
-                          "You might need to restore it manually", self.machine.label, e)
-
-            # after all this, we can make the Resultserver forget about it
-            Resultserver().del_task(self.task, self.machine)
+                          "You might need to restore it manually",
+                          self.machine.label, e)
 
         return succeeded
 
@@ -269,26 +331,59 @@ class AnalysisManager(Thread):
         # If the target is a file and the user enabled the option,
         # delete the original copy.
         if self.task.category == "file" and self.cfg.cuckoo.delete_original:
-            try:
-                os.remove(self.task.target)
-            except OSError as e:
-                log.error("Unable to delete original file at path \"%s\": %s", self.task.target, e)
+            if not os.path.exists(self.task.target):
+                log.warning("Original file does not exist anymore: \"%s\": "
+                            "File not found", self.task.target)
+            else:
+                try:
+                    os.remove(self.task.target)
+                except OSError as e:
+                    log.error("Unable to delete original file at path "
+                              "\"%s\": %s", self.task.target, e)
 
-        log.info("Task #%d: reports generation completed (path=%s)", self.task.id, self.storage)
+        # If the target is a file and the user enabled the delete copy of
+        # the binary option, then delete the copy.
+        if self.task.category == "file" and self.cfg.cuckoo.delete_bin_copy:
+            if not os.path.exists(self.binary):
+                log.warning("Copy of the original file does not exist anymore: \"%s\": File not found", self.binary)
+            else:
+                try:
+                    os.remove(self.binary)
+                except OSError as e:
+                    log.error("Unable to delete the copy of the original file at path \"%s\": %s", self.binary, e)
+
+        log.info("Task #%d: reports generation completed (path=%s)",
+                 self.task.id, self.storage)
 
         return True
 
     def run(self):
         """Run manager thread."""
-        success = self.launch_analysis()
-        Database().set_status(self.task.id, TASK_COMPLETED)
+        global active_analysis_count
+        active_analysis_count += 1
+        try:
+            while True:
+                try:
+                    success = self.launch_analysis()
+                except CuckooDeadMachine:
+                    continue
 
-        log.debug("Released database task #%d with status %s", self.task.id, success)
+                break
 
-        self.process_results()
-        Database().set_status(self.task.id, TASK_REPORTED)
+            Database().set_status(self.task.id, TASK_COMPLETED)
 
-        log.info("Task #%d: analysis procedure completed", self.task.id)
+            log.debug("Released database task #%d with status %s",
+                      self.task.id, success)
+
+            if self.cfg.cuckoo.process_results:
+                self.process_results()
+                Database().set_status(self.task.id, TASK_REPORTED)
+
+            log.info("Task #%d: analysis procedure completed", self.task.id)
+        except:
+            log.exception("Failure in AnalysisManager.run")
+
+        active_analysis_count -= 1
 
 class Scheduler:
     """Tasks Scheduler.
@@ -310,7 +405,7 @@ class Scheduler:
         """Initialize the machine manager."""
         global machinery
 
-        machinery_name = self.cfg.cuckoo.machine_manager
+        machinery_name = self.cfg.cuckoo.machinery
 
         log.info("Using \"%s\" machine manager", machinery_name)
 
@@ -325,14 +420,17 @@ class Scheduler:
 
         if not os.path.exists(conf):
             raise CuckooCriticalError("The configuration file for machine "
-                                      "manager \"{0}\" does not exist at path: "
-                                      "{1}".format(machinery_name, conf))
+                                      "manager \"{0}\" does not exist at path:"
+                                      " {1}".format(machinery_name, conf))
 
         # Provide a dictionary with the configuration options to the
         # machine manager instance.
         machinery.set_options(Config(conf))
         # Initialize the machine manager.
-        machinery.initialize(machinery_name)
+        try:
+            machinery.initialize(machinery_name)
+        except CuckooMachineError as e:
+            raise CuckooCriticalError("Error initializing machines: %s" % e)
 
         # At this point all the available machines should have been identified
         # and added to the list. If none were found, Cuckoo needs to abort the
@@ -350,6 +448,7 @@ class Scheduler:
 
     def start(self):
         """Start scheduler."""
+        global total_analysis_count
         self.initialize()
 
         log.info("Waiting for analysis tasks...")
@@ -357,25 +456,55 @@ class Scheduler:
         # Message queue with threads to transmit exceptions (used as IPC).
         errors = Queue.Queue()
 
+        maxcount = self.cfg.cuckoo.max_analysis_count
+
         # This loop runs forever.
         while self.running:
             time.sleep(1)
+
+            # If not enough free diskspace is available, then we print an
+            # error message and wait another round (this check is ignored
+            # when freespace is set to zero).
+            if self.cfg.cuckoo.freespace:
+                # Resolve the full base path to the analysis folder, just in
+                # case somebody decides to make a symlink out of it.
+                dir_path = os.path.join(CUCKOO_ROOT, "storage", "analyses")
+
+                # TODO: Windows support
+                if hasattr(os, "statvfs"):
+                    dir_stats = os.statvfs(dir_path)
+
+                    # Free diskspace in megabytes.
+                    space_available = dir_stats.f_bavail * dir_stats.f_frsize
+                    space_available /= 1024 * 1024
+
+                    if space_available < self.cfg.cuckoo.freespace:
+                        log.error("Not enough free diskspace! (Only %d MB!)",
+                                  space_available)
+                        continue
 
             # If no machines are available, it's pointless to fetch for
             # pending tasks. Loop over.
             if machinery.availables() == 0:
                 continue
 
-            # Fetch a pending analysis task.
-            task = self.db.fetch()
+            # Exits if max_analysis_count is defined in config file and
+            # is reached.
+            if maxcount and total_analysis_count >= maxcount:
+                if active_analysis_count <= 0:
+                    self.stop()
+            else:
+                # Fetch a pending analysis task.
+                task = self.db.fetch()
 
-            if task:
-                log.debug("Processing task #%s", task.id)
+                if task:
+                    log.debug("Processing task #%s", task.id)
+                    total_analysis_count += 1
 
-                # Initialize the analysis manager.
-                analysis = AnalysisManager(task, errors)
-                # Start.
-                analysis.start()
+                    # Initialize the analysis manager.
+                    analysis = AnalysisManager(task, errors)
+                    # Start.
+                    analysis.start()
 
             # Deal with errors.
             try:
